Improved gstreamer-sharp audio backend
Should be fully functional now, I still need to check the difference between stopping and closing a stream. I think close will free all resources and finalize the stream, whilst stop is supposed to set the position to 0 so that the stream can be recovered. But in this case FadeAndStop should be renamed to FadeAndClose because currently in all cases where FadeAndStop is called the stream should be closed instead of being stopped.