Fix Lost Profile Bug in Player Selection
Non-guest profiles assigned to player numbers higher than current amount
of players were left as already selected and thus couldn't be selected
anymore.

Now the player numbers that are higher than the totals will be cleared
and the assigned profiles put back to the pool of profiles to select
from.

New profiles will have to be selected if the player amount is increased.