Show song name text on lyric position top if possible
Until now there was no song name text displayed, if lyrics position is set to top or a duet song was startet.
This change check, if there is enough time before the first lyrics will be displayed and show the song name text, if there is enough time. If necessary, the max display time of 10 seconds will be shortened down to a value between 10 to 4 seconds.