Key shortcut to show only songs of the current playlist
CTRL+L while a playlist is open. It's a toggle key, just press again to return to normal list.
While filtered to a playlist, you can also use CTRL+L while the playlist is not open to switch back to the normal view.

My initial plans included also a button on the playlist pane, to the left of the playlist save button. I was able to add this button and set events on it. But I don't know how to filter and refresh the song screen, while I'm in the CPlaylist class of the VocaluxeLib namespace. After playing around for hours I stopped trying this and added just this key shortcut.