Read R (RAP) and G (Golden RAP) notes as Freestyle notes
Read R (RAP) and G (Golden RAP) notes as Freestyle notes currently. Final implementation needed!

As mentioned in #479 this fix the song loading of songs, which contains R or G notes. But Vocaluxe still don't support these kind of notes, so the final implementation is needed. If someone wants to work on it, feel free.